small updates

+++ app/Http/Controllers/ProductionController.php
@@ -45,8 +45,9 @@ class ProductionController extends Controller implements HasMiddleware
     public function index($type)
     {
         $materials = $this->materials;
+        $products = Material::where('type', 'parts')->orderBy('id')->get();
         $productions = Production::withTrashed()->where('type', $type)->where('branch_id', Session::get('branch'))->latest()->get();
-        return view('production.index', compact('productions', 'materials', 'type'));
+        return view('production.index', compact('productions', 'materials', 'products', 'type'));
     }
 
     /**
@@ -84,6 +85,9 @@ class ProductionController extends Controller implements HasMiddleware
                             'production_id' => $production->id,
                             'material_id' => $item,
                             'qty' => $request->qty[$key],
+                            'type' => 'out',
+                            'created_by' => $request->user()->id,
+                            'updated_by' => $request->user()->id,
                             'created_at' => Carbon::now(),
                             'updated_at' => Carbon::now(),
                         ];
@@ -110,9 +114,17 @@ class ProductionController extends Controller implements HasMiddleware
      */
     public function edit(string $type, string $id)
     {
-        $production = Production::findOrFail(decrypt($id));
-        $materials = $this->materials;
-        $companies = $this->company;
+        try {
+            $production = Production::findOrFail(decrypt($id));
+            if ($production->details()->where('type', 'in')->exists()):
+                throw new Exception("User not allowed to edit this record at the moment");
+            endif;
+            $materials = $this->materials;
+            $companies = $this->company;
+        } catch (Exception $e) {
+            return redirect()->back()->with("error", $e->getMessage());
+        }
+
         return view('production.edit', compact('production', 'materials', 'companies', 'type'));
     }
 
@@ -140,12 +152,15 @@ class ProductionController extends Controller implements HasMiddleware
                             'production_id' => $production->id,
                             'material_id' => $item,
                             'qty' => $request->qty[$key],
+                            'type' => 'out',
+                            'created_by' => $request->user()->id,
+                            'updated_by' => $request->user()->id,
                             'created_at' => $production->created_at,
                             'updated_at' => Carbon::now(),
                         ];
                     endif;
                 endforeach;
-                ProductionDetails::where('production_id', $production->id)->delete();
+                ProductionDetails::where('production_id', $production->id)->where('type', 'out')->delete();
                 ProductionDetails::insert($data);
             });
         } catch (Exception $e) {
@@ -159,16 +174,23 @@ class ProductionController extends Controller implements HasMiddleware
      */
     public function destroy(string $type, string $id)
     {
-        $production = Production::where('id', decrypt($id))->firstOrFail();
-        $production->delete();
-        ProductionDetails::where('production_id', decrypt($id))->whereNull('deleted_at')->delete();
+        try {
+            $production = Production::where('id', decrypt($id))->firstOrFail();
+            if ($production->details()->where('type', 'in')->exists()):
+                throw new Exception("User not allowed to delete this record at the moment");
+            endif;
+            $production->delete();
+            ProductionDetails::where('production_id', decrypt($id))->where('type', 'out')->whereNull('deleted_at')->delete();
+        } catch (Exception $e) {
+            return redirect()->back()->with("error", $e->getMessage());
+        }
         return redirect()->route('production.register', $type)->with("success", "Production deleted successfully");
     }
 
     public function restore(string $type, string $id)
     {
         $production = Production::withTrashed()->where('id', decrypt($id))->first();
-        ProductionDetails::withTrashed()->where('production_id', decrypt($id))->where('deleted_at', $production->deleted_at)->restore();
+        ProductionDetails::withTrashed()->where('production_id', decrypt($id))->where('type', 'out')->where('deleted_at', $production->deleted_at)->restore();
         $production->restore();
         return redirect()->route('production.register', $type)->with("success", "Production restored successfully");
     }
